+++ toolium/test/test_visual_test.py
@@ -237,6 +237,22 @@ class VisualTests(unittest.TestCase):
         # Assert output image
         self.assert_image(img, self._testMethodName, 'register')
 
+    def test_assert_screenshot_no_enabled_force(self):
+        # Configure driver mock
+        def copy_file_side_effect(output_file):
+            shutil.copyfile(self.file_v1, output_file)
+
+        self.driver_wrapper.driver.save_screenshot.side_effect = copy_file_side_effect
+
+        # Update conf and create a new VisualTest instance
+        self.driver_wrapper.config.set('VisualTests', 'enabled', 'false')
+        self.visual = VisualTest(self.driver_wrapper, force=True)
+
+        # Assert screenshot
+        self.visual.assert_screenshot(None, filename='screenshot_full', file_suffix='screenshot_suffix')
+        output_file = os.path.join(self.visual.output_directory, '01_screenshot_full__screenshot_suffix.png')
+        self.driver_wrapper.driver.save_screenshot.assert_called_with(output_file)
+
     def test_assert_screenshot_full_and_save_baseline(self):
         # Configure driver mock
         def copy_file_side_effect(output_file):